+++ lib/thump.dart
@@ -1,6 +1,12 @@
 import 'dart:collection';
 import 'dart:math';
 
+class Vector2 {
+  final double x;
+  final double y;
+  Vector2(this.x, this.y);
+}
+
 /// An axis-aligned bounding box.
 class AABB {
   /// The lowest x value for this.
@@ -112,12 +118,46 @@ class AABBPair {
   AABBPair(this.object, this.aabb);
 }
 
+class Collision {
+  /// The object that was collided with.
+  final Object object;
+
+  /// The [AABB] of the object that was collided with.
+  final AABB aabb;
+
+  /// The edge of the moving [AABB] that collided with [object].
+  final Edge edge;
+
+  /// The behavior executed at the collision.
+  final Behavior behavior;
+
+  /// The position of the object's top left hand corner at the point of impact.
+  final Vector2 collisionPosition;
+
+  Collision({
+    required this.object,
+    required this.aabb,
+    required this.edge,
+    required this.behavior,
+    required this.collisionPosition,
+  });
+}
+
 /// A result from [World.move].
 class MoveResult {
-  final double x;
-  final double y;
-  final List<AABBPair> collisions;
-  MoveResult(this.x, this.y, this.collisions);
+  /// The resulting position of the moved object.
+  final Vector2 position;
+
+  /// The resulting normalized direction the object is traveling.
+  final Vector2 direction;
+
+  final List<Collision> collisions;
+
+  MoveResult._make({
+    required this.position,
+    required this.collisions,
+    required this.direction,
+  });
 }
 
 class _Node {
@@ -130,36 +170,36 @@ class _Node {
   Map<Object, AABB> _entries = {};
 }
 
-enum _Edge { top, right, bottom, left }
+enum Edge { top, right, bottom, left }
 
-_Edge _calcClosestEdge(AABB a, AABB b, double dx, double dy) {
-  _Edge? result;
+Edge _calcClosestEdge(AABB a, AABB b, double dx, double dy) {
+  Edge? result;
   double minDist = double.infinity;
   if (dx >= 0) {
     double dist = (a.right - b.x).abs();
     if (dist < minDist) {
-      result = _Edge.right;
+      result = Edge.right;
       minDist = dist;
     }
   }
   if (dx <= 0) {
     double dist = (a.x - b.right).abs();
     if (dist < minDist) {
-      result = _Edge.left;
+      result = Edge.left;
       minDist = dist;
     }
   }
   if (dy >= 0) {
     double dist = (a.bottom - b.y).abs();
     if (dist < minDist) {
-      result = _Edge.bottom;
+      result = Edge.bottom;
       minDist = dist;
     }
   }
   if (dy <= 0) {
     double dist = (a.y - b.bottom).abs();
     if (dist < minDist) {
-      result = _Edge.top;
+      result = Edge.top;
       minDist = dist;
     }
   }
@@ -343,7 +383,8 @@ class World {
     final AABB start = _aabbs[obj]!;
     double resultX = start.x;
     double resultY = start.y;
-    Set<Object> collisions = {};
+    Set<Object> collisionObjects = {};
+    List<Collision> collisions = [];
     final AABB end = AABB.xywh(
         x: start.x + dx,
         y: start.y + dy,
@@ -351,6 +392,9 @@ class World {
         height: start.height);
     final AABB union = start.union(end);
     final List<AABBPair> potentials = queryAABB(union, ignore: obj);
+    final length = _length(dx, dy);
+    final double origDirX = dx / length;
+    final double origDirY = dy / length;
     if (potentials.isEmpty) {
       resultX = start.x + dx;
       resultY = start.y + dy;
@@ -361,9 +405,12 @@ class World {
               y: resultY,
               width: start.width,
               height: start.height));
-      return MoveResult(resultX, resultY, []);
+      return MoveResult._make(
+          position: Vector2(resultX, resultY),
+          collisions: [],
+          direction: Vector2(origDirX, origDirY));
     }
-    final int steps = _length(dx, dy).ceil();
+    final int steps = length.ceil();
     double normDx = dx / steps;
     double normDy = dy / steps;
     bool shouldBreak = false;
@@ -374,81 +421,79 @@ class World {
           x: nextX, y: nextY, width: start.width, height: start.height);
       for (AABBPair potential in potentials) {
         if (nextAABB.overlaps(potential.aabb)) {
-          collisions.add(potential.object);
+          final Edge closest = _calcClosestEdge(
+              AABB.xywh(
+                  x: resultX,
+                  y: resultY,
+                  width: start.width,
+                  height: start.height),
+              potential.aabb,
+              dx,
+              dy);
           Behavior behavior = handler(potential.object);
+          late final Vector2 collisionPosition;
+          switch (closest) {
+            case Edge.top:
+              final double y = max(nextY, potential.aabb.bottom);
+              double moveRatio = (y - resultY) / normDy;
+              final double x = resultX + normDx * moveRatio;
+              collisionPosition = Vector2(x, y);
+            case Edge.right:
+              final double x = min(nextX, potential.aabb.x - start.width);
+              double moveRatio = (x - resultX) / normDx;
+              final double y = resultY + normDy * moveRatio;
+              collisionPosition = Vector2(x, y);
+            case Edge.bottom:
+              final double y = min(nextY, potential.aabb.y - start.height);
+              double moveRatio = (y - resultY) / normDy;
+              final double x = resultX + normDx * moveRatio;
+              collisionPosition = Vector2(x, y);
+            case Edge.left:
+              final double x = max(nextX, potential.aabb.right);
+              double moveRatio = (x - resultX) / normDx;
+              final double y = resultY + normDy * moveRatio;
+              collisionPosition = Vector2(x, y);
+          }
+
+          if (!collisionObjects.contains(potential.object)) {
+            collisions.add(Collision(
+                object: potential.object,
+                aabb: potential.aabb,
+                edge: closest,
+                behavior: behavior,
+                collisionPosition: collisionPosition));
+            collisionObjects.add(potential.object);
+          }
           switch (behavior) {
             case Behavior.Touch:
               shouldBreak = true;
-              final _Edge closest = _calcClosestEdge(
-                  AABB.xywh(
-                      x: resultX,
-                      y: resultY,
-                      width: start.width,
-                      height: start.height),
-                  potential.aabb,
-                  dx,
-                  dy);
-              switch (closest) {
-                case _Edge.top:
-                  nextY = max(nextY, potential.aabb.bottom);
-                  double moveRatio = (nextY - start.y) / normDy;
-                  nextX = resultX + normDx * moveRatio;
-                case _Edge.right:
-                  nextX = min(nextX, potential.aabb.x - start.width);
-                  double moveRatio = (nextX - start.x) / normDx;
-                  nextY = resultY + normDy * moveRatio;
-                case _Edge.bottom:
-                  nextY = min(nextY, potential.aabb.y - start.height);
-                  double moveRatio = (nextY - start.y) / normDy;
-                  nextX = resultX + normDx * moveRatio;
-                case _Edge.left:
-                  nextX = max(nextX, potential.aabb.right);
-                  double moveRatio = (nextX - start.x) / normDx;
-                  nextY = resultY + normDy * moveRatio;
-              }
+              nextY = collisionPosition.y;
+              nextX = collisionPosition.x;
               break;
             case Behavior.Slide:
-              final _Edge closest = _calcClosestEdge(
-                  AABB.xywh(
-                      x: resultX,
-                      y: resultY,
-                      width: start.width,
-                      height: start.height),
-                  potential.aabb,
-                  dx,
-                  dy);
               switch (closest) {
-                case _Edge.top:
+                case Edge.top:
                   nextY = max(nextY, potential.aabb.bottom);
-                case _Edge.right:
+                case Edge.right:
                   nextX = min(nextX, potential.aabb.x - start.width);
-                case _Edge.bottom:
+                case Edge.bottom:
                   nextY = min(nextY, potential.aabb.y - start.height);
-                case _Edge.left:
+                case Edge.left:
                   nextX = max(nextX, potential.aabb.right);
               }
               break;
             case Behavior.Pass:
               break;
             case Behavior.Bounce:
-              final _Edge closest = _calcClosestEdge(
-                  AABB.xywh(
-                      x: resultX,
-                      y: resultY,
-                      width: start.width,
-                      height: start.height),
-                  potential.aabb,
-                  dx,
-                  dy);
               switch (closest) {
-                case _Edge.top:
+                case Edge.top:
                   final double impactY = max(nextY, potential.aabb.bottom);
                   final double moveRatio = (impactY - resultY) / normDy;
                   nextY = resultY +
                       normDy * moveRatio +
                       (-normDy * (1 - moveRatio));
                   normDy *= -1;
-                case _Edge.right:
+                case Edge.right:
                   final double impactX =
                       min(nextX, potential.aabb.x - start.width);
                   final double moveRatio = (impactX - resultX) / normDx;
@@ -456,7 +501,7 @@ class World {
                       normDx * moveRatio +
                       (-normDx * (1 - moveRatio));
                   normDx *= -1;
-                case _Edge.bottom:
+                case Edge.bottom:
                   final double impactY =
                       min(nextY, potential.aabb.y - start.height);
                   final double moveRatio = (impactY - resultY) / normDy;
@@ -464,7 +509,7 @@ class World {
                       normDy * moveRatio +
                       (-normDy * (1 - moveRatio));
                   normDy *= -1;
-                case _Edge.left:
+                case Edge.left:
                   final double impactX = max(nextX, potential.aabb.right);
                   final double moveRatio = (impactX - resultX) / normDx;
                   nextX = resultX +
@@ -486,7 +531,11 @@ class World {
         AABB.xywh(
             x: resultX, y: resultY, width: start.width, height: start.height));
 
-    return MoveResult(resultX, resultY,
-        collisions.map((Object obj) => AABBPair(obj, _aabbs[obj]!)).toList());
+    return MoveResult._make(
+      position: Vector2(resultX, resultY),
+      collisions: collisions,
+      direction: Vector2((normDx < 0 != dx < 0) ? -origDirX : origDirX,
+          (normDy < 0 != dy < 0) ? -origDirY : origDirY),
+    );
   }
 }