+++ test/thump_test.dart
@@ -1,5 +1,9 @@
+import 'dart:math';
+
 import 'package:test/test.dart';
 import 'package:thump/thump.dart';
+
+final double sqrt2Reciprocal = 1.0 / sqrt(2.0);
 
 void main() {
   test('queryObject', () {
@@ -85,10 +89,13 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 47, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 0);
-    expect(result.x, closeTo(48.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(48.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].edge, Edge.right);
+    expect(result.collisions[0].collisionPosition.x, closeTo(48.0, 0.01));
+    expect(result.collisions[0].collisionPosition.y, closeTo(16.0, 0.01));
   });
 
   test('simple slide right tiny', () {
@@ -98,8 +105,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 1, height: 1));
     world.add(man, AABB.xywh(x: 61.99, y: 16, width: 1, height: 1));
     MoveResult result = world.move(man, 2.99, 0);
-    expect(result.x, closeTo(63.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(63.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -126,10 +133,11 @@ void main() {
     world.add(block, AABB.xywh(x: 48, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 48, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 3, 4);
-    expect(result.x, closeTo(51.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(51.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].edge, Edge.bottom);
   });
 
   test('simple slide down right fractional', () {
@@ -139,8 +147,8 @@ void main() {
     world.add(block, AABB.xywh(x: 48, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 48, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 3, 3);
-    expect(result.x, closeTo(51, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(51, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -152,8 +160,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 47, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 0, handler: (obj) => Behavior.Pass);
-    expect(result.x, closeTo(49.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(49.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -165,10 +173,11 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 81, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, -2, 0);
-    expect(result.x, closeTo(80.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(80.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].edge, Edge.left);
   });
 
   test('simple slide up', () {
@@ -178,10 +187,12 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 64, y: 33, width: 16, height: 16));
     MoveResult result = world.move(man, 0, -2);
-    expect(result.x, closeTo(64.0, 0.01));
-    expect(result.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(64.0, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].edge, Edge.top);
+    expect(result.collisions[0].behavior, Behavior.Slide);
   });
 
   test('slide slip by', () {
@@ -191,8 +202,8 @@ void main() {
     world.add(block, AABB.xywh(x: 16, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 0, y: 50, width: 16, height: 16));
     MoveResult result = world.move(man, 0, -50);
-    expect(result.x, closeTo(0.0, 0.01));
-    expect(result.y, closeTo(0.0, 0.01));
+    expect(result.position.x, closeTo(0.0, 0.01));
+    expect(result.position.y, closeTo(0.0, 0.01));
     expect(result.collisions.length, 0);
   });
 
@@ -203,8 +214,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 64, y: 15, width: 16, height: 16));
     MoveResult result = world.move(man, 0, 2);
-    expect(result.x, closeTo(64.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(64.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
 
@@ -226,8 +237,8 @@ void main() {
     world.add(man, AABB.xywh(x: 47.9, y: 16, width: 16, height: 16));
     MoveResult result =
         world.move(man, 10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.x, closeTo(48.0, 0.01));
-    expect(result.y, closeTo(16.1, 0.01));
+    expect(result.position.x, closeTo(48.0, 0.01));
+    expect(result.position.y, closeTo(16.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -240,8 +251,8 @@ void main() {
     world.add(man, AABB.xywh(x: 80.1, y: 16, width: 16, height: 16));
     MoveResult result =
         world.move(man, -10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.x, closeTo(80.0, 0.01));
-    expect(result.y, closeTo(16.1, 0.01));
+    expect(result.position.x, closeTo(80.0, 0.01));
+    expect(result.position.y, closeTo(16.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -254,8 +265,8 @@ void main() {
     world.add(man, AABB.xywh(x: 64, y: 32.1, width: 16, height: 16));
     MoveResult result =
         world.move(man, 10, -10, handler: (obj) => Behavior.Touch);
-    expect(result.y, closeTo(32.0, 0.01));
-    expect(result.x, closeTo(64.1, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(64.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -268,8 +279,8 @@ void main() {
     world.add(man, AABB.xywh(x: 64, y: 15.9, width: 16, height: 16));
     MoveResult result =
         world.move(man, -10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.y, closeTo(16.0, 0.01));
-    expect(result.x, closeTo(63.9, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(63.9, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -283,8 +294,8 @@ void main() {
     world.add(blockDown, AABB.xywh(x: 32, y: 48, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 31, y: 31, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 2);
-    expect(result.y, closeTo(32.0, 0.01));
-    expect(result.x, closeTo(32.0, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(32.0, 0.01));
     expect(result.collisions.length, 2);
   });
 
@@ -296,10 +307,14 @@ void main() {
     world.add(man, AABB.xywh(x: 48, y: 14, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(52.0, 0.01));
-    expect(result.y, closeTo(14.0, 0.01));
+    expect(result.position.x, closeTo(52.0, 0.01));
+    expect(result.position.y, closeTo(14.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(-sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].collisionPosition.x, closeTo(50.0, 0.01));
+    expect(result.collisions[0].collisionPosition.y, closeTo(16.0, 0.01));
   });
 
   test('simple bounce top', () {
@@ -310,8 +325,10 @@ void main() {
     world.add(man, AABB.xywh(x: 48, y: 50, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, -4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(52.0, 0.01));
-    expect(result.y, closeTo(50.0, 0.01));
+    expect(result.position.x, closeTo(52.0, 0.01));
+    expect(result.position.y, closeTo(50.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -324,8 +341,10 @@ void main() {
     world.add(man, AABB.xywh(x: 30, y: 32, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(30.0, 0.01));
-    expect(result.y, closeTo(36.0, 0.01));
+    expect(result.position.x, closeTo(30.0, 0.01));
+    expect(result.position.y, closeTo(36.0, 0.01));
+    expect(result.direction.x, closeTo(-sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -338,9 +357,12 @@ void main() {
     world.add(man, AABB.xywh(x: 66, y: 32, width: 16, height: 16));
     MoveResult result =
         world.move(man, -4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(66.0, 0.01));
-    expect(result.y, closeTo(36.0, 0.01));
+    expect(result.position.x, closeTo(66.0, 0.01));
+    expect(result.position.y, closeTo(36.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
+    expect(result.collisions[0].behavior, Behavior.Bounce);
   });
 }